Combine Streamable and Processable, remove unused events.
